feat: update to loam
