+++ crates/smartdeploy/Cargo.toml
@@ -1,5 +1,7 @@
 [package]
 name = "smartdeploy"
+description = "A crate for managing and deploying smart contracts on the Soroban blockchain."
+documentation = "https://docs.rs/smartdeploy"
 version = "0.0.0"
 authors = ["Stellar Development Foundation <info@stellar.org>"]
 license = "Apache-2.0"
